Merges `tests/test_dependency.py` with `tests/dependency_injection/test_dependency.py`
